requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
pre-commit
pydantic-settings
python-multipart
aiofiles
pillow
pytesseract
